move dto and descrambler into separate files
